isp_ticketing_streamlit_pioneer_pg/requirements.txt
streamlit>=1.38.0
SQLAlchemy>=2.0.32
pandas>=2.2.2
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
